requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
pillow>=10.3.0
openai==1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pytest==7.4.3
httpx==0.25.2
pydantic[email]
